+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   instance methods. This would result in a substantial memory leak during batch processing because the surface objects
   would not be garbage collected once they went out of scope. 
 - Added cache to the height paramters, reducing calculation time by ~20% since Sq is invoked multiple times. 
+- Fixed dtype mismatch in Cython code for topographies with float32 instead of float64 dtype (usually plu and plux 
+  files) 
 ## v0.4.0:
 - Reduced file load times by >90% using numpy.fromfile to read the height data
 - Leveling now no longer also centers the data around the mean